comment unnecessary methods
